認証のSpringSecurity適用完了

+++ backend/src/main/java/com/daily_app/demo/Repository/RemindRepository.java
@@ -18,6 +18,6 @@ public interface RemindRepository extends JpaRepository<Remind, Integer>{
     // 「毎日リマインドを送りたい（毎日レコードを作りたい）」のであれば、存在チェックに「今日の日付」も含める必要がある
     //boolean existsByUserAndRemindContentAndCreatedAtAfter(User user, String remindContent, LocalDateTime dateTime);
 
-    List<Remind> findByUser_UserId(Long userId);
+    List<Remind> findByUser_UserId(Integer userId);
     
 }
